chore: extract `getNavigationElement` from `get-breadcrumbs-for-element.ts`
